+++ components/main/bottom/Bottom.tsx
@@ -1,6 +1,6 @@
 /* eslint-disable @next/next/no-img-element */
 import Link from "next/link"
-import Container from "./Container"
+import Container from "../../shared/Container"
 
 export const Bottom = () => {
     return (
